setup-env: Allow override of GITHUB_USER and DOCKER_REPOSITORY
It's sometimes useful to specify values for GITHUB_USER and
DOCKER_REPOSITORY that are not derived from TRAVIS_REPO_SLUG.  Two
examples are when not building with Travis or when the github user is
different from the Docker repository name.  DOCKER_USERNAME can be set
in the Travis settings just like the password.