Be nicer to the official repository
By using a local mirror of the official repository we preserve its resources. This is done by making
git://git.sv.gnu.org point to a local mirror that is kept up to date. For HTTPS, simply socat to the real HTTPS
server (only the master images make use of the HEAD to bust their cache).